requirements.txt
discord-py-interactions>=4.2.0
python-dotenv~=0.19.2
aiohttp~=3.8.1